gcp/requirements.txt
<<<<<<< HEAD
tensorflow==2.12.0
google-cloud-storage==2.8.0
Pillow==8.4.0
=======
tensorflow==2.12
google-cloud-storage==2.9.0
Pillow==9.0.1
>>>>>>> origin/main
